requirement.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.1.0
EasyProcess==1.1
entrypoint2==1.1
ffmpeg-python==0.2.0
future==0.18.3
iniconfig==2.0.0
jproperties==2.1.1
mss==9.0.1
names==0.3.0
natsort==8.4.0
numpy==1.25.2
opencv-python==4.8.0.76
packaging==23.1
Pillow==10.0.0
pluggy==1.2.0
pyscreenrec==0.4
pyscreenshot==3.1
PyScreeze==0.1.29
pytest==7.4.0
six==1.16.0
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.1.0
certifi==2023.7.22
EasyProcess==1.1
entrypoint2==1.1
exceptiongroup==1.1.3
ffmpeg-python==0.2.0
future==0.18.3
h11==0.14.0
idna==3.4
iniconfig==2.0.0
jproperties==2.1.1
mss==9.0.1
names==0.3.0
natsort==8.4.0
numpy==1.25.2
opencv-python==4.8.0.76
outcome==1.2.0
packaging==23.1
Pillow==10.0.0
pluggy==1.2.0
pyscreenrec==0.4
pyscreenshot==3.1
PyScreeze==0.1.29
PySocks==1.7.1
pytest==7.4.0
selenium==4.11.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.2
trio-websocket==0.10.3
urllib3==2.0.4
wsproto==1.2.0
